+++ forui/lib/src/widgets/avatar/avatar_content.dart
@@ -14,22 +14,30 @@ class const Content({
 }) extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final fallback = this.fallback ?? PlaceholderContent(style: style, size: size);
+    final fallback = FittedBox(
+      fit: .scaleDown,
+      child: this.fallback ?? PlaceholderContent(style: style, size: size),
+    );
+
     final duration = context.accessibility.motion == .disabled
         ? Duration.zero
         : style(context.theme.avatarStyle).fadeInDuration;
 
-    return Image(
-      height: size,
-      width: size,
-      image: image,
-      semanticLabel: semanticsLabel,
-      errorBuilder: (_, _, _) => fallback,
-      frameBuilder: (_, child, frame, wasSynchronouslyLoaded) => wasSynchronouslyLoaded
-          ? child
-          : AnimatedSwitcher(duration: duration, child: frame == null ? fallback : child),
-      loadingBuilder: (_, child, loadingProgress) => loadingProgress == null ? child : fallback,
-      fit: BoxFit.cover,
+    return Semantics(
+      label: semanticsLabel,
+      image: true,
+      excludeSemantics: true,
+      child: Image(
+        height: size,
+        width: size,
+        image: image,
+        errorBuilder: (_, _, _) => fallback,
+        frameBuilder: (_, child, frame, wasSynchronouslyLoaded) => wasSynchronouslyLoaded
+            ? child
+            : AnimatedSwitcher(duration: duration, child: frame == null ? fallback : child),
+        loadingBuilder: (_, child, loadingProgress) => loadingProgress == null ? child : fallback,
+        fit: .cover,
+      ),
     );
   }