+++ forui/lib/src/widgets/select/single/select.dart
@@ -919,7 +919,15 @@ abstract class _State<S extends FSelect<T>, T> extends State<S> with TickerProvi
             ),
           ),
           child: CallbackShortcuts(
-            bindings: {const SingleActivator(.enter): _toggle, const SingleActivator(.space): _toggle},
+            bindings: {
+              const SingleActivator(.enter): _toggle,
+              const SingleActivator(.space): _toggle,
+              const SingleActivator(.backspace): () {
+                if (widget.enabled && widget.clearable && _focus.hasPrimaryFocus) {
+                  _textController.clear();
+                }
+              },
+            },
             child: ListenableBuilder(
               listenable: _popoverController,
               child: widget.builder(context, style, variants, field),